Split import function into separate functions
Split the logToPlan function into 2 functions, import and transform for ease of use. i.e one function for importing, one function for transforming (smart importing will need this anyway so we can have both functions and one import function)
disabled unholy nova from importing (filter in future etc)
