+++ src/General/Modules/CooldownPlanner/Engine/SmartTransformData.js
@@ -1,0 +1,1 @@
+// Function will create plan data by merging healer casts that are used near boss abilities, rather than create a new instance for each seperate cast.
